feat(cb2-14394): refactor handler file
